refactor - modify order tables package structure
- 주문 테이블과 관련된 패키지 구조를 에그리거트 중심 구조로 변경했습니다
