+++ src/main/java/kitchenpos/orders/bo/OrderBo.java
@@ -3,11 +3,11 @@ package kitchenpos.orders.bo;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.orders.dao.OrderDao;
 import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.ordertables.dao.OrderTableDao;
 import kitchenpos.orders.model.Order;
 import kitchenpos.model.OrderLineItem;
 import kitchenpos.orders.model.OrderStatus;
-import kitchenpos.model.OrderTable;
+import kitchenpos.ordertables.model.OrderTable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
