+++ src/test/java/kitchenpos/bo/OrderBoTest.java
@@ -3,7 +3,7 @@ package kitchenpos.bo;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.orders.dao.OrderDao;
 import kitchenpos.dao.OrderLineItemDao;
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.ordertables.dao.OrderTableDao;
 import kitchenpos.orders.model.Order;
 import kitchenpos.orders.model.OrderStatus;
 import kitchenpos.orders.bo.OrderBo;
